requirements.txt
customtkinter>=5.2.0
tkinterdnd2>=0.3.0
openai-whisper>=20231117
